Move meeting types to module.
